Install shellcheck and shfmt for deploy workflow
Both of the release scripts call `pnpm verify:bail` (which runs the test
suite), so these test dependencies need to be installed here as well.

Although `shellcheck` is included in GitHub's `ubuntu-latest` build
image, we explicitly install it here both for consistency with the
`verify` workflow and as a guard against this being removed from the
build image in future.