Add support for --simplify shfmt option
This wasn't included initially as it does more than just format code.
However, its absence has been noted in a comment on the last PR:
https://github.com/bash-lsp/bash-language-server/pull/1136#issuecomment-2090804329
and it is a valid `shfmt` option, so support has been added.

There's a similar option (`-mn` or `--minify`), but this will not be
implemented as its output is not intended for human consumption and it
will essentially ignore all other options presented anyway.