Add support for reading shfmt config options from .editorconfig
If we have any `shfmt`-specific options in `.editorconfig`, use the
config in `.editorconfig` and ignore the language server config (this is
similar to `shfmt`'s approach of using either `.editorconfig` or command
line flags, but not both).

Indentation always comes via the editor - if someone is using
`.editorconfig` then the expectation is that they will have configured
their editor's indentation in this way too.