shellcheck, shebang: rework shell type detection
Perform a more rigorous shell type (dialect) detection to avoid
interfering with shellcheck. In particular, track whether we detected
a shell dialect from the shebang, `shellcheck shell=...` directive or
the file name (URI). When invoking shellcheck, only pass an override
(i.e., the `--shell=` argument) or bail out on an unsupported dialect
if we guessed something that is not in the document text.

Otherwise, do not pass _anything_ and let shellcheck work it out on its
own because it's smarter than us.

This makes sure that
1) the user can always override our logic with a `shell=` directive;
2) we handle unknown shell types as gracefully as possible.